requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Pillow
gunicorn
